pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
version = "1.6.1"
name = "maelzel"
description = "Framework for computer music and composition"
authors = [
    { name = "Eduardo Moguillansky", email = "eduardo.moguillansky@gmail.com" },
]
license = { file = "LICENSE.md" }
readme = "README.rst"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    'Topic :: Multimedia :: Sound/Audio',
]


dependencies = [
    "numpy",
    "scipy",
    "matplotlib",
    "appdirs",
    "tabulate",
    "cachetools",
    "pyyaml",
    "psutil",
    "lxml",
    "quicktions",
    "rich",
    "chardet",
    "simple-term-menu",
    "visvalingamwyatt",
    "distro",
    "thefuzz",
    # "resampy",  # Do not include it by default, since this pulls numba
    "python-constraint",
    "pillow<11",
    "sounddevice",
    "typing_extensions",
    "pypng",
    # Own libraries
    "emlib>=1.18.0",
    "libcsound>=0.10.6",
    "bpf4>=1.10.0",
    "configdict>=2.10",
    "sndfileio>=1.9.1",
    "numpyx>=1.3.1",
    "csoundengine>=2.16.0",
    "pitchtools>=1.14.0",
    "risset>=3.2.0",
    "loristrck>=1.6.1",
    "vamphost>=1.3.0",
    "lilyponddist>=1.0.1",
    "jupyterlab>=4.3.5",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["maelzel", "maelzel.*"]
exclude = ["docs", "notebooks", "test", "scripts"]


[tool.setuptools.package-data]
maelzel = ['maelzel/data']


[project.urls]
"Homepage" = "https://github.com/gesellkammer/maelzel"


[tool.ruff.lint]
# 2. Avoid enforcing line-length violations (`E501`)
ignore = ["E501", "I001", "E402"]
